# `gix fetch` with fast-forward support