… README.md
… git-pack/tests/pack/bundle.rs
… git-protocol/tests/fetch/response.rs
… git-protocol/tests/fetch/v1.rs
… git-protocol/tests/fixtures/v1/clone-with-keepalive.response
… git-protocol/tests/fixtures/v2/clone-only-with-keepalive.response
… git-repository/src/remote/connection/fetch/update_refs/tests.rs
… git-repository/tests/remote/fetch.rs
… git-repository/tests/remote/ref_map.rs
… git-traverse/src/commit.rs
… git-traverse/tests/commit/mod.rs
… git-traverse/tests/fixtures/generated-archives/make_traversal_repo_for_commits_with_dates.tar.xz
… gitoxide-core/src/repository/fetch.rs
… gitoxide-core/src/repository/mod.rs
… gitoxide-core/src/repository/remote.rs
… src/plumbing/main.rs
… src/plumbing/options/mod.rs
… tests/journey/gix.sh
… tests/snapshots/plumbing/no-repo/pack/index/create/output-dir-content
… tests/snapshots/plumbing/no-repo/pack/receive/file-v-any-with-output
… tests/snapshots/plumbing/no-repo/pack/receive/ls-in-output-dir
… tests/snapshots/plumbing/repository/remote/refs/remote ref-list-no-networking-in-small-failure