* [x] fast determination of fast-forward, with the use of dates in the currently loaded 'level' nodes that will be iterated to determine a stop condition: if everything is already older than the commit we look for, stop.
* ~~use dry-run `fetch` in `gix remote refs` to provide more detailed information about what would be done to refs~~ - it's OK to do as is I think, as fetch currently also does negotiation even in dry-run mode. Maybe that's something to avoid, maybe it's something that can help later. For now it will simply avoid receiving the pack entirely.
* [x] `gix fetch` to perform an actual fetch, with dry-run mode
* [x] a fetch-test that validates the output filenames - needs `pack-` prefix
* [x] an issue with progress reporting blocking the clone the first time you try (if the clone is fast enough)
* [x] deal with keep-alive lines which are empty data lines before the pack is set and only relevant in V1 which may not send progress before the pack is sent.
* [x] correctly set protocol version on spawned git process


### Performance Issues

* [x] `git_tempfile` uses a NamedTempFile under the hood, and that's not buffered. Probably should be wrapped into a buf-writer to assure decent write performance. Right now, 60% of CPU time go into writing a tempfile. BufWriter should be around it in the bundle writer.

### Performance Notes

* fetch the whole linux kernel in ~53s with 8.5 cores, git does that in 3m with 3 cores
   * with 1 core, 2m 25s, and git takes 3m 47s
   * with 3 cores like git: 1m 17s
   * we use 1.4GB peak memory, git takes 3.05GB

### Out of scope

- async implementation
- worktree updates
- merges

### Previous PR

* #539